corregir tipo de dato matrices

+++ taller_cuda/main_program.c
@@ -4,7 +4,7 @@
 #include <inttypes.h>
 #include <string.h>
 
-void filtro(uint8_t* matriz_entrada, int rows, int cols, uint8_t* matriz_salida){
+void filtro(int16_t* matriz_entrada, int rows, int cols, int16_t* matriz_salida){
     // implementar el filtro como corresponde 
     // esto es solo dummy 
     memcpy(matriz_salida, matriz_entrada, rows*cols);
@@ -29,8 +29,8 @@ int main(int argc, char** argv){
     if(fscanf(input, "%d\n", &rows)==EOF) return 3;
     if(fscanf(input, "%d\n", &cols) ==EOF) return 3;
     int pixel_count = rows*cols;
-    uint8_t* src_image = (uint8_t*) malloc(pixel_count*sizeof(uint8_t));
-    uint8_t* out_image = (uint8_t*) malloc(pixel_count*sizeof(uint8_t));
+    int16_t* src_image = (int16_t*) malloc(pixel_count*sizeof(int16_t));
+    int16_t* out_image = (int16_t*) malloc(pixel_count*sizeof(int16_t));
     if(src_image == NULL) return 3;
     int i = 0;
     int pixel;
